requirements.txt
chainlit
openai-agents
autogen-agentchat
autogen-ext[mcp]
autogen-ext[openai]
autogen-ext[ollama]
